+++ src/pages/dashboard/index.js
@@ -3,11 +3,11 @@ export * from "@/pages/dashboard/accounts";
 export * from "@/pages/dashboard/activities";
 export * from "@/pages/dashboard/contacts";
 export * from "@/pages/dashboard/opportunities";
+export * from "@/pages/dashboard/products";
 export * from "@/pages/dashboard/reports";
 export * from "@/pages/dashboard/maps"
 export * from "@/pages/dashboard/prospects";
 export * from "@/pages/dashboard/admin";
-
 
 // export * from "@/pages/dashboard/profile";
 // export * from "@/pages/dashboard/tables";